Move CacheAllocation and KVCache to seperate file since they are not paged_attention specific.
