+++ sharktank/sharktank/layers/__init__.py
@@ -6,7 +6,8 @@
 
 from .base import *
 from .conv import Conv2DLayer, Conv3DLayer, Conv1DLayer
-from .paged_attention import PagedAttention, attn_type_map, CacheAllocation, KVCache
+from .attention import CacheAllocation, KVCache
+from .paged_attention import PagedAttention, attn_type_map
 from .causal_llm import BaseCausalLMModel
 from .linear import LinearLayer
 from .norm import RMSNormLayer, LayerNorm
